pyproject.toml
[project]
name = "allgoodpy"
version = "0.1.0"
description = ""
authors = [
    {name = "dilweb",email = "yxs.crevier@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.15,<4",
    "aiohttp>=3.9",
    "opencv-python-headless>=4.10,<5",
    "numpy>=1.26,<3",
    "pyzbar>=0.1.9",
    "pillow>=10,<12",
    "pytesseract>=0.3.10",
    "gspread>=6.0,<7",
    "google-auth>=2.29,<3",
    "pydantic-settings>=2.6,<3",
]

[project.scripts]
allgoodpy-bot = "allgoodpy.bot.main:main"

[tool.poetry]
packages = [{include = "allgoodpy", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
